128x128/requirements_interface.txt
# Core dependencies for the Car GAN Web Interface
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
